requirements.txt
django>1.5
django-ckeditor-updated==4.2.8
